requirements (2).txt
tensorflow
numpy
matplotlib
pillow
scikit-learn
streamlit